ref(discover) - Moving handle_data to api base
- This will allow the Key Transactions endpoint to use the same
  handle_data functions
